Create generic `ActionContextMenuViewModifier`
